+++ docs/whats-new.rst
@@ -176,7 +176,7 @@ Changes and Deprecations
 
         Cleanup options have been split into ``cleanup`` and
         ``cleanup_on_failure``. There are more granular values for both of
-        these options. More info: :doc:`configs-runners`
+        these options.
 
         Most limitations have been lifted from passthrough options, including
         the former inability to use custom types and actions. More info: