+++ docs/whats-new.rst
@@ -4,6 +4,17 @@ What's New
 For a complete list of changes, see `CHANGES.txt
 <https://github.com/Yelp/mrjob/blob/master/CHANGES.txt>`_
 
+0.3.5
+-----
+
+The *pool_wait_minutes* (:opt:`--pool-wait-minutes`) option lets your job delay
+itself in case a job flow becomes available. Reference:
+:doc:`configs-reference`
+
+The ``JOB`` and ``JOB_FLOW`` cleanup options tell mrjob to clean up the job
+and/or the job flow on failure (including Ctrl+C). See
+:py:data:`~mrjob.runner.CLEANUP_CHOICES` for more information.
+
 0.3.3
 -----